requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.0
psycopg2-binary==2.9.7
python-multipart==0.0.6
pytest==7.4.2
httpx==0.24.1